Some cleanup around prompts / credentials
